Update Redis cache server edgi-govdata-archiving/web-monitoring#119
Our previous cache was in us-east-1, which is very far away from where our services are now (they used to be on Heroku, which was in us-east-1). I've set up a new Redis instance in us-west-2 and incremented the `api` and `import-worker` configs here so they pick up the new value (it's in the secrets file, which you can find in Keybase).