Add incident report for 2-month Wayback Failure (#16)
It seems I left the checkout of web-monitoring-processing that loads data in production from the Wayback Machine in a bad, merge-conflict state, so it failed immediately with Python syntax errors every time it ran. No process was set up to alert for this situation. (We normally rely on Sentry for this, but it runs in-process and if Python never actually runs our code, it doesn’t get a chance to work.)

The problem persisted for nearly two months (since about the start of the shutdown).

See also this explanation from the Dev meeting the next day: https://youtu.be/vcpmvMppM-0?t=654