Describe RDS Configuration in “Manually Managed” Section (#26)
Yesterday I set up a new production database in RDS, and spent a fair amount of time poking around to try and get it configured as well as I reasonably could. (A new one was easier because so much needed to change from the current production DB.) This PR adds information about how our RDS databases are configured.

The new database is named `web-monitoring-db-production-a`, and I’ll remove the old one once we’ve been running comfortable on this new one for a week.