Set Sentry environment names for diffing server
I just triggered a bunch of imports to staging, which caused some errors based on old, bad data we had in that database. That's totally fine (part of the reason we have staging!), but I discovered we aren't differentiating staging and production for the diffing server!

In the -db project, we use totally different Sentry projects for staging and production. They have an environment feature, though, so I'm giving that a try with the diffing server. We might consider switching to it for DB in the future if it works well.