Document ETL server setup (#28)
We've never clearly documented how the sever that moves data from the Wayback Machine to our database and that generates weekly analyst sheets works, and that's a real problem! This lays out how things currently operate.

(That's separate from the fact that the current setup is kind of a mess! It could use some improvement, but at least the current state of affairs is documented now.)