Reorganize Documentation and Write a Real README (#27)
The main goal here is to replace the README with something that actually gives an overview of the repo and how to approach it. The current README has been moved to a separate folder with Kubernetes setup information. (I haven’t attempted to change that tutorial at all otherwise.)

This also renames the templates directory to kubernetes (hopefully that makes things more clear) and shoves the old README and associated setup files/scripts in kubernetes/0-setup-guide. The kubernetes/README.md file still needs to be written.

This is a part of edgi-govdata-archiving/web-monitoring#138.