Update page statuses on a regular schedule (#29)
We now have tasks we need to run on a regular schedule — a job needs to go through and update the `status` field on Page objects in web-monitoring-db every so often. (They have to be updated on a schedule rather than based on a particular event because the results are based on a time window — the status could change just by time passing without new versions coming in.)

This adds Kubernetes `CronJob` configurations, which are basically containers that can be run on a cron-style schedule. At current, this runs them twice a day (every 12 hours). I already deployed this to staging to test, and it works great.

Fixes edgi-govdata-archiving/web-monitoring#122.