Add Redis cache to staging
We apparently weren't using any caching in staging, which caused me to think that a test deploy was fine, even though it turned out to fail when I deployed it to the production namespace. Staging now uses the same Redis cache server as production (but a different database).

Associated with be02e0380a525bfed7fe32f8c16e5822562850b1 in the ops-internal repo.